% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ARIEvolution}
\alias{ARIEvolution}
\title{Plot the evolution of the pairwise ARIs as merging
happens}
\usage{
ARIEvolution(
  merger,
  unclustered = NULL,
  values = TRUE,
  numericalLabels = FALSE,
  state_length = 1
)
}
\arguments{
\item{merger}{the result from having run \code{\link{Dune}}
on the dataset}

\item{unclustered}{The value assigned to unclustered cells. Default to
\code{NULL}}

\item{values}{Whether to also display the ARI values. Default to TRUE.}

\item{numericalLabels}{Whether labels are numerical values. Default to FALSE.}

\item{state_length}{Time between steps. Default to 1. See \code{\link{transition_states}}
for details.}
}
\value{
a \code{gganim} object
}
\description{
Plot the evolution of the pairwise ARIs as merging
happens
}
\details{
See \code{\link{plotARIs}} and \code{\link{animate}}.
}
\examples{
\dontrun{
  data("clusMat", package = "Dune")
  merger <- Dune(clusMat = clusMat)
  ARIEvolution(merger)}
}
