% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ARIImp}
\alias{ARIImp}
\title{ARI improvement}
\usage{
ARIImp(merger)
}
\arguments{
\item{merger}{the result from having run \code{\link{mergeManyPairwise}}
on the dataset}
}
\value{
a vector with the mean ARI between methods at each step
}
\description{
Compute the ARI improvement over the ARI merging procedure
}
