% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ARItrend}
\alias{ARItrend}
\title{ARI improvement plot}
\usage{
ARItrend(merger)
}
\arguments{
\item{merger}{the result from having run \code{\link{mergeManyPairwise}}
on the dataset}
}
\value{
a \code{\link{ggplot}} object
}
\description{
A plot to see how ARI improves over merging
}
