% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{assignRsec}
\alias{assignRsec}
\title{Assign cells using the assignUnassigned function of RSEC}
\usage{
assignRsec(merger, p = 1)
}
\arguments{
\item{merger}{the result from having run \code{\link{Dune}}
on the dataset}

\item{p}{when to stop the merging, when mean ARI has improved to p (between 0
and 1) of the final value.}
}
\description{
By default, RSEC does not assign all cells, leaving those as "-1".
However, to give a fair comparison with other labels, it is necessary to
assign cells so it is necessary to track that along the merging
}
