% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{clusterMatToAri}
\alias{clusterMatToAri}
\title{Compute an ARI matrix and plot an heatmap of it}
\usage{
clusterMatToAri(mat, small = FALSE)
}
\arguments{
\item{mat}{the matrix of clustering labels (one per columns)}

\item{small}{whether to also display the values}
}
\value{
a \code{\link{ggplot}} object
}
\description{
We can compute the ARI between pairs of cluster labels. This function plots
a matrix where a cell is the adjusted Rand Index between cluster label of
row i and cluster label of column j.
}
