% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ConfusionPlot}
\alias{ConfusionPlot}
\title{Plot confusion matrix}
\usage{
ConfusionPlot(x, y = NULL)
}
\arguments{
\item{x}{A vector of clustering labels or a matrix of clustering labels. See details.}

\item{y}{Optional. Another vector of clustering labels}
}
\value{
a \code{\link{ggplot}} object
}
\description{
A plot to visualize how alike two clustering labels are
}
\examples{
data("nuclei", package = "Dune")
ConfusionPlot(clusMat[, c("SC3", "Monocle")])
}
