% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{Consensus}
\alias{Consensus}
\title{Find the consensus clustering between three methods and return the consensus}
\usage{
Consensus(clusMat, large = FALSE, ...)
}
\arguments{
\item{clusMat}{The clustering matrix with a row per cell and a column per
clustering label type}

\item{large}{If the dataset is too large to be handled by \code{\link{MakeConsensus}},
we use a different method to get the consensus clustering. Default to FALSE.}

\item{...}{Other arguments passed to \code{\link{MakeConsensus}}}
}
\value{
a vector of cluster assignations for the consensus.
}
\description{
Find the consensus clustering between three methods and return the consensus
}
