% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.adjustedRandIndex}
\alias{.adjustedRandIndex}
\title{adjustedRandIndex}
\usage{
.adjustedRandIndex(tab)
}
\arguments{
\item{table}{The confusion matrix}
}
\value{
The ARI
}
\description{
adjustedRandIndex
}
\details{
Second part of the \code{\link{adjustedRandIndex}} function from
\code{\link{mclust}}.
}
