% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.dupRemove}
\alias{.dupRemove}
\title{Remove the duplicated cells which belong to the same cluster over all labels.}
\usage{
.dupRemove(clusMat)
}
\arguments{
\item{clusMat}{The clustering matrix with a row per cell and a column per
clustering label type}
}
\value{
a list containing the reduced matrix and a way to map back to the original
}
\description{
Remove the duplicated cells which belong to the same cluster over all labels.
}
