% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dune.R
\name{Dune}
\alias{Dune}
\title{Perform the ARI merging}
\usage{
Dune(clusMat, unclustered = NULL, nCores = 1, verbose = FALSE)
}
\arguments{
\item{clusMat}{the matrix of samples by clustering labels.}

\item{unclustered}{The value assigned to unclustered cells. Default to \code{NULL}}

\item{nCores}{number of cores to use when parallelizing. Relies on
\code{\link{mclapply}}. Default to 1.}

\item{verbose}{Whether or not the print cluster merging as it happens.}
}
\value{
A list with four components: the initial matrix of clustering labels,
the final matrix of clustering labels, the merge info matrix and the ARI
improvement vector.
}
\description{
Compute the ARI between every pair of clustering labels after merging every possible pair of clusters. Find the one that improves the ARI merging the most, merge the pair. Repeat until there is no improvement.
}
\details{
The Dune algorithm merges pairs of clusters in order to improve the
mean adjusted Rand Index with other clustering labels. It returns a list with
four components.:
#' \itemize{
\item \code{initialMat}: The initial matrix of cluster labels
\item \code{currentMat}: The final matrix of cluster labels
\item \code{merges}: The step-by-step detail of the merges, recapitulating
which clusters where merged in which cluster label
\item \code{impARI}: How much each merge improved the mean ARI between the
cluster label that has been merged and the other cluster labels.
}
}
\examples{
data("clusMat", package = "Dune")
merger <- Dune(clusMat = clusMat)
# clusters 11 to 14 from cluster label 5 and 3 are subset of cluster 2 from
# other cluster labels. Designing cluster 2 as unclustered stops the merge of those.
merger2 <- Dune(clusMat = clusMat, unclustered = 2)
merger$merges
merger2$merges
}
\seealso{
clusterConversion ARIImp
}
