% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{FTracking}
\alias{FTracking}
\title{Track the evolution of a function along merging}
\usage{
FTracking(merger, f, ...)
}
\arguments{
\item{merger}{the result from having run \code{\link{Dune}}
on the dataset}

\item{f}{the function used, can be computed on any partition of the space}

\item{...}{additional arguments passed to f}
}
\value{
a matrix with a column per initial clustering, and a row per merge
with the f value computed
}
\description{
For a given ARI merging, compute the evolution on the function f with
another partition
}
