% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{mergeManyPairwise}
\alias{mergeManyPairwise}
\title{Perform the ARI merging}
\usage{
mergeManyPairwise(clusteringMatrix, nCores = 3)
}
\arguments{
\item{clusteringMatrix}{the matrix of samples by clustering labels.}

\item{nCores}{number of cores to use when parallelizing. Relies on \code{\link{mclapply}}.}
}
\value{
a list with the initial matrix of clustering labels, the final matrix of clustering labels, the merge info matrix and the ARI improvement.
}
\description{
Compute the ARI between every pair of clustering labels after merging every possible pair of clusters. Find the one that improves the ARI merging the most, merge the pair. Repeat until there is no improvement.
}
