% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotARIReduce}
\alias{plotARIReduce}
\title{plot the ARI improvement between methods}
\usage{
plotARIReduce(merger, RSEC = FALSE)
}
\arguments{
\item{merger}{the result from having run \code{\link{mergeManyPairwise}}
on the dataset}

\item{RSEC}{default to \code{FALSE}. Whether there is an RSEC cluster labels
(i.e one that does not cluster all cells)}
}
\value{
the output from \code{\link{plot_grid}}
}
\description{
The output from this function is a grid of 4 plots, based on the
\code{\link{plotARIs}} function. The first row of 2 plots is before the
merging procedure, the second is after the merging procedure. The first
column is when RSEC uses all assigned cells, the second column is when only
using the cells that RSEC cluster for computing the ARI between RSEC and
another partition of the data.
}
