% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{whenToStop}
\alias{whenToStop}
\title{ARI Matrix}
\usage{
whenToStop(merger, p)
}
\arguments{
\item{merger}{the result from having run \code{\link{Dune}}
on the dataset}

\item{p}{A value between 0 and 1. We stop when the mean ARI has improved by p
of the final total improvement. Default to 1 (i.e running the full merging).}
}
\value{
An integer giving the step where to stop.
}
\description{
ARI Matrix
}
\details{
The \code{\link{Dune}} process improves the mean ARI. This return
the first merging step after which the mean ARI has been improved by p of the
total. Setting p = 1 just return the number of merges.
}
\examples{
data("clusMat", package = "Dune")
merger <- Dune(clusMat = clusMat)
whenToStop(merger, p = .5)
}
