% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ARItrend}
\alias{ARItrend}
\title{ARI improvement plot}
\usage{
ARItrend(merger, unclustered = NULL)
}
\arguments{
\item{merger}{the result from having run \code{\link{Dune}}
on the dataset}

\item{unclustered}{The value assigned to unclustered cells. Default to
\code{NULL}}
}
\value{
a \code{\link{ggplot}} object
}
\description{
A plot to see how ARI improves over merging
}
\examples{
data("clusMat", package = "Dune")
merger <- Dune(clusMat = clusMat)
ARItrend(merger)
}
