% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{clusterConversion}
\alias{clusterConversion}
\title{clusterConversion}
\usage{
clusterConversion(merger, unclustered = NULL)
}
\arguments{
\item{merger}{the result from having run \code{\link{Dune}}
on the dataset}

\item{unclustered}{The value assigned to unclustered cells. Default to \code{NULL}}
}
\value{
a vector with the mean ARI between methods at each merge
}
\description{
Find the conversion between the old cluster and the final clusters
}
