% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{intermediateMat}
\alias{intermediateMat}
\title{Find the clustering matrix that we would get if we stopped the ARI merging
early}
\usage{
intermediateMat(merger, p = 1, n_steps = NULL)
}
\arguments{
\item{merger}{the result from having run \code{\link{Dune}}
on the dataset}

\item{p}{A value between 0 and 1. We stop when the mean ARI has improved by p
of the final total improvement. Default to 1 (i.e running the full merging).}

\item{n_steps}{Alternatively, you can specifiy the number of merging steps to
do before stopping.}
}
\value{
A matrix with the same dimensions as the currentmMat of the merger
argument
}
\description{
Find the clustering matrix that we would get if we stopped the ARI merging
early
}
\examples{
data("clusMat", package = "Dune")
merger <- Dune(clusMat = clusMat)
head(intermediateMat(merger, n_steps = 1))
}
