% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{intermediateMat}
\alias{intermediateMat}
\title{Find the clustering matrix that we would get if we stopped the ARI merging
early}
\usage{
intermediateMat(merger, p = 0.9)
}
\arguments{
\item{merger}{the result from having run \code{\link{mergeManyPairwise}}
on the dataset}

\item{p}{A value between 0 and 1. We stop when the mean ARI has improved by p
of the final total improvement}
}
\value{
A matrix with the same dimensions as the currentmMat of the merger
argument
}
\description{
Find the clustering matrix that we would get if we stopped the ARI merging
early
}
