% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nuclei}
\alias{nuclei}
\title{Cluster labels for a subset of the allen Smart-Seq nuclei dataset}
\format{An object of class \code{data.frame} with 1744 rows and 7 columns.}
\usage{
nuclei
}
\description{
Cluster labels for a subset of the allen Smart-Seq nuclei dataset
}
\details{
This matrix of clusters was obtained by running all those algo then cutting it down
to the GABAergic neurons. See...
}
\keyword{datasets}
