% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotARIs}
\alias{plotARIs}
\title{Plot an heatmap of the ARI matrix}
\usage{
plotARIs(ARI, small = T)
}
\arguments{
\item{ARI}{the matrix of pairwise ARI}

\item{small}{whether to also display the values}
}
\value{
a \code{\link{ggplot}} object
}
\description{
We can compute the ARI between pairs of cluster labels. This function plots
a matrix where a cell is the adjusted Rand Index between cluster label of
row i and cluster label of column j.
}
