% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotPrePost}
\alias{plotPrePost}
\title{Plot the reduction in cluster size for an ARI merging}
\usage{
plotPrePost(merger)
}
\arguments{
\item{merger}{The output from an ARI merging}
}
\value{
a \code{\link{ggplot}} object
#' @importFrom dplyr mutate
}
\description{
Plot the reduction in cluster size for an ARI merging
}
